Add Decision Tree Classifier
Added a Random Forest in an attempt to compare and contrast between the two. May change later this week for something else but reckon there's enough here to play off of to get a decent grade. The main focus right now is to optimise and get the best out of both since the code itself isn't graded.